backend/requirements.txt
fastapi==0.104.1
sqlmodel==0.0.16
pydantic==2.5.0
pydantic-settings==2.1.0
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
pyjwt==2.8.0
python-multipart==0.0.6
better-exceptions==0.3.3
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
sqlalchemy==2.0.23
alembic==1.12.1
python-dotenv==1.0.0
bcrypt==4.0.1